Skip test TestAccProjectResource
Ubicloud's API can now be accessed with a personal access token, which
has advantages over the user-based token received from the login API.
One limitation of personal access tokens is that they are project scoped.
Creating new projects fails with a 403 when authenticated with a
personal access token, which is expected. As `TestAccProjectResource`
creates/destroys project it is disabled by default so that the whole test
suite passes with personal access tokens. `TestAccProjectResource`
can still be manually run as necessary.